misc clang-tidy cleanup

+++ src/core/time.h
@@ -8,12 +8,12 @@
 namespace barely {
 
 [[nodiscard]] constexpr double FramesToSeconds(double sample_rate, int64_t frames) noexcept {
-  assert(sample_rate > 0);
-  return static_cast<double>(frames) / static_cast<double>(sample_rate);
+  assert(sample_rate > 0.0);
+  return static_cast<double>(frames) / sample_rate;
 }
 
 [[nodiscard]] constexpr int64_t SecondsToFrames(double sample_rate, double seconds) noexcept {
-  return static_cast<int64_t>(seconds * static_cast<double>(sample_rate));
+  return static_cast<int64_t>(seconds * sample_rate);
 }
 
 [[nodiscard]] constexpr double Quantize(double position, int32_t subdivision,
